requirements.txt
h5py==3.10.0
matplotlib==3.8.1
numpy==1.26.2
pandas==2.1.3
PyYAML==6.0.1
#ray==2.8.0
scikit_learn==1.3.2
setuptools==68.2.2
tensorflow==2.15.0
typer==0.9.0
